# server/openapi/v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Streamarr Core Server API",
    "description": "Interface-agnostic API for Usenet search, resolve, and byte-range streaming (BRIEF §6.2).",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/config/general": {
      "get": {
        "tags": [
          "GeneralConfig"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralConfigResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralConfigResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralConfigResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "GeneralConfig"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneralConfigWrite"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneralConfigWrite"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GeneralConfigWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralConfigResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralConfigResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralConfigResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "host": {
            "type": "string",
            "description": "Populated only for the `nzb_host_not_allowed` error: the rejected download host.",
            "nullable": true
          },
          "indexerId": {
            "type": "string",
            "description": "Populated only for the `nzb_host_not_allowed` error: the owning indexer's id.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ErrorResponse": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "description": "Typed error envelope rendered consistently by every endpoint."
      },
...
      "GeneralConfigResponse": {
        "type": "object",
        "properties": {
          "tmdbApiKey": {
            "type": "string",
            "nullable": true
          },
          "hasTmdbApiKey": {
            "type": "boolean"
          },
          "sessionTtlSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "ephemeralCacheSizeMb": {
            "type": "integer",
            "format": "int32"
          },
          "searchCacheTtlSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "indexerResultLimit": {
            "type": "integer",
            "format": "int32"
          },
          "segmentCacheSizeMb": {
            "type": "integer",
            "format": "int32"
          },
          "connectionBudget": {
            "type": "integer",
            "format": "int32"
          },
          "addStreamarrBadge": {
            "type": "boolean"
          },
          "addReleaseScoreToName": {
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "description": "General config as returned by the config API — the TMDB key is masked."
      },
      "GeneralConfigWrite": {
        "type": "object",
        "properties": {
          "tmdbApiKey": {
            "type": "string",
            "nullable": true
          },
          "sessionTtlSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ephemeralCacheSizeMb": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "searchCacheTtlSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "indexerResultLimit": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "segmentCacheSizeMb": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "connectionBudget": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "addStreamarrBadge": {
            "type": "boolean",
            "nullable": true
          },
          "addReleaseScoreToName": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Write model for general config. Secret Streamarr.Server.Config.GeneralConfigWrite.TmdbApiKey is write-only."
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "Machine API key or admin session JWT.",
        "scheme": "bearer",
        "bearerFormat": "API key or JWT"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}